# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  healthy: 'healthy',
  low_stock: 'low_stock',
  out_of_stock: 'out_of_stock',
} as const;

export interface Product {
  id: number;
  sku: string;
  name: string;
  category: string;
  unitPrice: number;
  stockOnHand: number;
  reorderPoint: number;
  stockValue: number;
  status: ProductStatus;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  sku: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  unitPrice: number;
  /** @minimum 0 */
  stockOnHand: number;
  /** @minimum 0 */
  reorderPoint: number;
}

export interface ProductUpdate {
  /** @minLength 1 */
  sku?: string;
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  /** @minimum 0 */
  unitPrice?: number;
  /** @minimum 0 */
  reorderPoint?: number;
}

export type StockAdjustmentInputType = typeof StockAdjustmentInputType[keyof typeof StockAdjustmentInputType];


export const StockAdjustmentInputType = {
  in: 'in',
  out: 'out',
  adjustment: 'adjustment',
} as const;

export interface StockAdjustmentInput {
  type: StockAdjustmentInputType;
  /** @minimum 1 */
  quantity: number;
  /** @nullable */
  note?: string | null;
}

export type InventoryMovementType = typeof InventoryMovementType[keyof typeof InventoryMovementType];


export const InventoryMovementType = {
  in: 'in',
  out: 'out',
  adjustment: 'adjustment',
} as const;

export interface InventoryMovement {
  id: number;
  productId: number;
  productName: string;
  type: InventoryMovementType;
  quantity: number;
  /** @nullable */
  note: string | null;
  createdAt: string;
}

export interface MovementTrendPoint {
  date: string;
  inbound: number;
  outbound: number;
}

export interface DashboardSummary {
  totalProducts: number;
  totalUnits: number;
  inventoryValue: number;
  lowStockCount: number;
  outOfStockCount: number;
  movementTrend: MovementTrendPoint[];
}

export interface SaleItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export type SaleInputPaymentMethod = typeof SaleInputPaymentMethod[keyof typeof SaleInputPaymentMethod];


export const SaleInputPaymentMethod = {
  cash: 'cash',
  card: 'card',
  upi: 'upi',
} as const;

export interface SaleInput {
  /** @minItems 1 */
  items: SaleItemInput[];
  paymentMethod: SaleInputPaymentMethod;
}

export interface SaleItem {
  productId: number;
  productName: string;
  sku: string;
  quantity: number;
  unitPrice: number;
  lineTotal: number;
}

export type SalePaymentMethod = typeof SalePaymentMethod[keyof typeof SalePaymentMethod];


export const SalePaymentMethod = {
  cash: 'cash',
  card: 'card',
  upi: 'upi',
} as const;

export interface Sale {
  id: number;
  saleNumber: string;
  items: SaleItem[];
  subtotal: number;
  total: number;
  paymentMethod: SalePaymentMethod;
  createdAt: string;
}

export type SaleSummaryPaymentMethod = typeof SaleSummaryPaymentMethod[keyof typeof SaleSummaryPaymentMethod];


export const SaleSummaryPaymentMethod = {
  cash: 'cash',
  card: 'card',
  upi: 'upi',
} as const;

export interface SaleSummary {
  id: number;
  saleNumber: string;
  itemCount: number;
  subtotal: number;
  total: number;
  paymentMethod: SaleSummaryPaymentMethod;
  createdAt: string;
}

export type ListProductsParams = {
search?: string;
category?: string;
status?: ListProductsStatus;
};

export type ListProductsStatus = typeof ListProductsStatus[keyof typeof ListProductsStatus];


export const ListProductsStatus = {
  all: 'all',
  healthy: 'healthy',
  low_stock: 'low_stock',
  out_of_stock: 'out_of_stock',
} as const;

export type ListMovementsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListSalesParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};