# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MovementTrendPoint } from './movementTrendPoint';

export interface DashboardSummary {
  totalProducts: number;
  totalUnits: number;
  inventoryValue: number;
  lowStockCount: number;
  outOfStockCount: number;
  movementTrend: MovementTrendPoint[];
}