# lib/api-zod/src/generated/types/movementTrendPoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MovementTrendPoint {
  date: string;
  inbound: number;
  outbound: number;
}