# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface Product {
  id: number;
  sku: string;
  name: string;
  category: string;
  unitPrice: number;
  stockOnHand: number;
  reorderPoint: number;
  stockValue: number;
  status: ProductStatus;
  updatedAt: Date;
}