# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  sku: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  unitPrice: number;
  /** @minimum 0 */
  stockOnHand: number;
  /** @minimum 0 */
  reorderPoint: number;
}