# lib/api-zod/src/generated/types/saleInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SaleInputPaymentMethod } from './saleInputPaymentMethod';
import type { SaleItemInput } from './saleItemInput';

export interface SaleInput {
  /** @minItems 1 */
  items: SaleItemInput[];
  paymentMethod: SaleInputPaymentMethod;
}