# lib/api-zod/src/generated/types/saleSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SaleSummaryPaymentMethod } from './saleSummaryPaymentMethod';

export interface SaleSummary {
  id: number;
  saleNumber: string;
  itemCount: number;
  subtotal: number;
  total: number;
  paymentMethod: SaleSummaryPaymentMethod;
  createdAt: Date;
}